+++ OsmAnd/res/values/strings.xml
@@ -11,7 +11,7 @@
 	Thx - Hardy
 
 -->
-    <string name="file_does_not_contains_routing_rules">\'%1$s\' file doesn\'t contain routing rules, please choose another file.</string>
+    <string name="file_does_not_contain_routing_rules">\'%1$s\' file doesn\'t contain routing rules, please choose another file.</string>
     <string name="not_support_file_type_with_ext">Not supported file type. You need to select a file with %1$s extension.</string>
     <string name="import_from_file">Import from file</string>
     <string name="import_routing_file">Import routing file</string>
